git-svn-id: http://svn.code.sf.net/p/jrat/code/trunk@116 51b2fdc4-721d-0410-97db-d1d35c42e7b1

+++ src/org/shiftone/jrat/core/Accumulator.java
@@ -131,12 +131,12 @@ public class Accumulator implements Serializable {
     }
 
 
-    public final Float getAverageDuration() {
+    public final Double getAverageDuration() {
 
-        Float average = null;
+        Double average = null;
 
         if (totalExits > 0) {
-            average = new Float((double) totalDuration / (double) totalExits);
+            average = new Double((double) totalDuration / (double) totalExits);
         }
 
         return average;
